+++ src/_templates/generator/help/index.ejs.t
