+++ src/_templates/generator/new/hello.ejs.t
