+++ src/_templates/generator/with-prompt/hello.ejs.t
