+++ src/_templates/generator/with-prompt/prompt.ejs.t
