+++ src/_templates/stories/new/stories.ejs.t
