+++ src/_templates/story/new/story.ejs.t
